+++ web/embed.go
@@ -4,5 +4,5 @@ import "embed"
 
 // FS contains embedded web assets from this directory.
 
-//go:embed auth-client.js mpr-sites.js
+//go:embed tauth.js mpr-sites.js
 var FS embed.FS